+++ src/main/kotlin/brokers/LocalBroker.kt
@@ -1,6 +1,7 @@
 package com.zamna.kotask
 
 import MDCContext
+import brokers.IMessageBroker
 import kotlinx.coroutines.*
 import kotlinx.coroutines.channels.Channel
 import kotlin.concurrent.thread
